backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "black>=25.9.0",
    "celery>=5.5.3",
    "channels>=4.3.1",
    "channels-redis>=4.3.0",
    "daphne>=4.2.1",
    "django>=5.2.8",
    "django-celery-beat>=2.8.1",
    "django-cors-headers>=4.9.0",
    "django-debug-toolbar>=6.1.0",
    "django-elasticsearch-dsl>=9.0",
    "django-extensions>=4.1",
    "django-filter>=25.2",
    "django-ratelimit>=4.1.0",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "factory-boy>=3.3.3",
    "faker>=37.12.0",
    "flake8>=7.3.0",
    "ipython>=9.7.0",
    "isort>=7.0.0",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.10.1",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.11.1",
    "pytest-mock>=3.15.1",
    "python-decouple>=3.8",
    "python-slugify>=8.0.4",
    "redis>=7.0.1",
    "sentry-sdk>=2.43.0",
]
